めちゃめちゃリファクタリング
