+++ .prettierrc.json
@@ -1,3 +1,3 @@
 {
-  "tabWidth": 4
+  "tabWidth": 2
 }
