+++ package.json
@@ -1,43 +1,43 @@
 {
-    "name": "return-youtube-comment-username",
-    "license": "MIT",
-    "version": "0.1.6",
-    "type": "commonjs",
-    "devDependencies": {
-        "@swc/core": "^1.3.36",
-        "@types/chrome": "^0.0.217",
-        "@types/express": "^4.17.17",
-        "@types/fs-extra": "^11.0.1",
-        "@types/glob": "^8.0.1",
-        "@types/node": "^18.14.0",
-        "@types/react": "^18.0.28",
-        "@types/sha256": "^0.2.0",
-        "@types/ws": "^8.5.4",
-        "@typescript-eslint/eslint-plugin": "^5.55.0",
-        "@typescript-eslint/parser": "^5.55.0",
-        "esbuild": "^0.17.10",
-        "esbuild-plugin-eslint": "^0.1.1",
-        "esbuild-register": "^3.4.2",
-        "eslint": "^8.31.0",
-        "eslint-config-prettier": "^8.6.0",
-        "eslint-plugin-react": "^7.31.11",
-        "eslint-plugin-react-hooks": "^4.6.0",
-        "express": "^4.18.2",
-        "fs-extra": "^11.1.0",
-        "glob": "^8.1.0",
-        "node-css-require": "^0.4.6",
-        "react-dom": "^18.2.0",
-        "sass": "^1.58.3",
-        "sha256": "^0.2.0",
-        "typescript": "^4.9.5",
-        "ws": "^8.12.1"
-    },
-    "dependencies": {},
-    "scripts": {
-        "esbuild-register": "node --require esbuild-register",
-        "build:userscript": "npm run esbuild-register build/ build mode=userScript",
-        "build:extension": "npm run esbuild-register build/ build mode=extension",
-        "build": "npm run build:userscript && npm run build:extension",
-        "dev": "npm run esbuild-register build/ dev mode=userScript"
-    }
+  "name": "return-youtube-comment-username",
+  "license": "MIT",
+  "version": "0.2.0",
+  "type": "commonjs",
+  "devDependencies": {
+    "@swc/core": "^1.3.36",
+    "@types/chrome": "^0.0.217",
+    "@types/express": "^4.17.17",
+    "@types/fs-extra": "^11.0.1",
+    "@types/glob": "^8.0.1",
+    "@types/node": "^18.14.0",
+    "@types/react": "^18.0.28",
+    "@types/sha256": "^0.2.0",
+    "@types/ws": "^8.5.4",
+    "@typescript-eslint/eslint-plugin": "^5.55.0",
+    "@typescript-eslint/parser": "^5.55.0",
+    "esbuild": "^0.17.10",
+    "esbuild-plugin-eslint": "^0.1.1",
+    "esbuild-register": "^3.4.2",
+    "eslint": "^8.31.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "express": "^4.18.2",
+    "fs-extra": "^11.1.0",
+    "glob": "^8.1.0",
+    "node-css-require": "^0.4.6",
+    "react-dom": "^18.2.0",
+    "sass": "^1.58.3",
+    "sha256": "^0.2.0",
+    "typescript": "^4.9.5",
+    "ws": "^8.12.1"
+  },
+  "dependencies": {},
+  "scripts": {
+    "esbuild-register": "node --require esbuild-register",
+    "build:userscript": "npm run esbuild-register build/ build mode=userScript",
+    "build:extension": "npm run esbuild-register build/ build mode=extension",
+    "build": "npm run build:userscript && npm run build:extension minify && npm run build:userscript minify",
+    "dev": "npm run esbuild-register build/ dev mode=userScript"
+  }
 }
