--- src/types.d.ts
@@ -1,4 +1,0 @@
-declare module "*.css" {
-    const classes: { [className: string]: string };
-    export = classes;
-}
